+++ src/utils.js
@@ -16,7 +16,6 @@ import {
   groupBy,
   map,
   isObject,
-  cloneDeep,
 } from 'lodash';
 import {
   itemStatusesMap,
@@ -181,6 +180,45 @@ export const validateAlphaNumericField = value => {
   }
 
   return undefined;
+};
+
+/**
+ * Provide validation of an optional form field consisting of one or more
+ * textfields and one or more select fields used for type id selection,
+ * where both parts (text and identifier type) are required.
+ *
+ * @param optionalField the field description object, consisting of
+ *  list: list name (string)
+ *  textFields: array of text field names
+ *  selectFields: array of select field names
+ * @param values array of field values passed in to caller validate function
+ *
+ * @return nested array of errors for optionalField
+ */
+export const validateOptionalField = (optionalField, values) => {
+  const listName = optionalField.list;
+  const errorList = [];
+
+  if (values[listName] && values[listName].length) {
+    values[listName].forEach((item, i) => {
+      const entryErrors = {};
+      optionalField.textFields.forEach((field) => {
+        if (!item || !item[field]) {
+          entryErrors[field] = <FormattedMessage id="ui-inventory.fillIn" />;
+          errorList[i] = entryErrors;
+        }
+      });
+
+      optionalField.selectFields.forEach((field) => {
+        if (!item || !item[field]) {
+          entryErrors[field] = <FormattedMessage id="ui-inventory.selectToContinue" />;
+          errorList[i] = entryErrors;
+        }
+      });
+    });
+  }
+
+  return errorList;
 };
 
 export const checkIfElementIsEmpty = element => ((!element || element === '-') ? noValue : element);